requirements.txt
opencv-python
pytesseract
Pillow
tqdm
python-dotenv
requests
yt-dlp
google-api-python-client
google-auth
google-auth-oauthlib
imageio
imageio-ffmpeg
moviepy
streamlink
setuptools
wheel
typing-extensions